[PORT FROM MAIN][R4] The wrong format ogg make the music player crash
BZ: 71305

In _vorbis_unpack_books function, when extracting
map backend settings, memory is allocated for the
mapping info structure. But the structure in itself
has pointers to chmuxlist, submaps, coupling_steps
which contains random addresses. So when mapping_info_unpack
returns error, mapping_clear_info is called to free the
memory if the pointer isn't NULL.
The current fix uses calloc instead of malloc to initialize
the mapping info structure to zero.
Signed-off-by: Manjunath <manjunathx.b.s@intel.com>

Change-Id: I4f0935f79b7e730c2098807aaf6511f7307207ff
origChange-Id: I80f391efc2eb522a426d76fb242f57119d265cc7
Reviewed-on: http://android.intel.com:8080/82058
Reviewed-by: M, Arulselvan <arulselvan.m@intel.com>
Tested-by: M, Arulselvan <arulselvan.m@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>